requirements.txt
absl-py==1.2.0
astor==0.8.1
Babel==2.10.3
backcall==0.2.0
bce-python-sdk==0.8.73
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.2
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.7.1
Flask==2.1.3
Flask-Babel==2.0.0
Flask-Cors==3.0.10
fonttools==4.34.4
future==0.18.2
google-auth==2.10.0
google-auth-oauthlib==0.4.6
grpcio==1.48.0
idna==3.3
imgviz==1.5.0
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver==1.4.4
labelme==5.0.1
lap==0.4.0
llvmlite==0.39.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.2.2
matplotlib-inline==0.1.3
motmetrics==1.2.5
natsort==8.1.0
nest-asyncio==1.5.5
numba==0.56.0
numpy==1.21.6
oauthlib==3.2.0
opencv-contrib-python-headless==4.6.0.66
opencv-python==4.1.1.26
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddlepaddle-gpu==2.3.1
paddleslim==2.2.1
paddlex==2.1.0
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycryptodome==3.15.0
Pygments==2.12.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
QtPy==2.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.23.2
scipy==1.7.3
seaborn==0.11.2
Shapely==1.8.2
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.10
visualdl==2.3.0
wcwidth==0.2.5
Werkzeug==2.1.2
xmltodict==0.13.0
zipp==3.8.1
